backend/requirements.txt
fastapi 
uvicorn[standard] 
aiogram
httpx
pydantic
pydantic-settings
python-dotenv